move testing helpers to feature:pdf
