--- core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeDateFormatter.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.testing
-
-import org.groundplatform.ui.util.DateFormatter
-
-/** [DateFormatter] for tests, so assertions don't depend on the host locale or time zone. */
-object FakeDateFormatter : DateFormatter {
-
-  override fun formatDate(millis: Long): String = "DATE($millis)"
-
-  override fun formatTime(millis: Long): String = "TIME($millis)"
-}
