--- core/testing/src/commonMain/kotlin/org/groundplatform/testing/FakeStringResolver.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2026 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.groundplatform.testing
-
-import org.groundplatform.ui.util.StringResolver
-import org.jetbrains.compose.resources.StringResource
-
-/**
- * [StringResolver] for tests so display logic can be asserted without a Compose resource runtime.
- */
-object FakeStringResolver : StringResolver {
-
-  override suspend fun resolve(resource: StringResource): String = resource.key
-
-  override suspend fun resolve(resource: StringResource, vararg formatArgs: Any): String =
-    "${resource.key}(${formatArgs.joinToString()})"
-}
